Stop calling fsync in CompilerFilter
`self.infile.flush()` is sufficient to ensure that the input is flushed
from the Python process to the filesystem.  We were also calling
`os.fsync(self.infile)`, which flushes the file to disk.  This is overly
conservative for a temporary file.

On my system this change saves ~50 ms per filter execution.