Fixed Django 1.8+ deprecation warnings for changed render_to_string/render
See https://docs.djangoproject.com/en/dev/releases/1.8/#dictionary-and-context-instance-arguments-of-rendering-functions

Note we are fixing both the changed render_to_string function and the fact
that Template.render doesn't want a 'Context' instance any more.

There is a slight backwards incompatible change: the post_compress signal
now receives a dictionary instead of a Context instance as its 'context'
argument.  However, this is compatible with what is publicly documented,
which specifies a 'context dictionary':

http://django-compressor.readthedocs.org/en/latest/usage/?highlight=context#compressor.signals.post_compress