Updating remote storages documentation for CachedS3BotoStorage bug.
The example in the docs doesn't quite work as the file stored in
the "content" variable is already closed, and ends up spitting out
encoded data to the destination files. It was properly being sent
to S3, but not getting saved locally.

This saves the file locally first, then using local storage, opens
the file once again for the S3BotoStorage save call.